+++ indexed_gzip/zran_file_util.pxd
@@ -16,11 +16,11 @@ cdef extern from "zran_file_util.h":
 
     size_t _fread_python(void *ptr, size_t size, size_t nmemb, PyObject *f)
     
-    long int _ftell_python(PyObject *f)
+    uint64_t _ftell_python(PyObject *f)
     
     int _fseek_python(PyObject *f, long int offset, int whence)
     
-    int _feof_python(PyObject *f, int64_t size)
+    int _feof_python(PyObject *f, uint64_t size)
     
     int _ferror_python(PyObject *f)
     
@@ -34,11 +34,11 @@ cdef extern from "zran_file_util.h":
 
     int fseek_(FILE *fd, PyObject *f, long int offset, int whence)
 
-    int ftell_(FILE *fd, PyObject *f)
+    uint64_t ftell_(FILE *fd, PyObject *f)
 
     size_t fread_(void *ptr, size_t size, size_t nmemb, FILE *fd, PyObject *f)
 
-    int feof_(FILE *fd, PyObject *f, int64_t size)
+    int feof_(FILE *fd, PyObject *f, uint64_t size)
 
     int fflush_(FILE *fd, PyObject *f)
